Restructuring, removing population-related files.
