+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "gulp test && istanbul cover _mocha tests",
     "coverage": "cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
-    "gen": "tsc && node dist/app.js"
+    "gen": "tsc && node dist/generator/app.js"
   },
   "repository": {
     "type": "git",
