+++ src/generator/progress.ts
